+++ apps/native-component-list/src/screens/Audio/expo-audio/Player.tsx
@@ -37,6 +37,7 @@ interface Props {
   pause: () => void;
   replay: () => void;
   next?: () => void;
+  replace?: () => void;
   setRate: (rate: number, shouldCorrectPitch: boolean) => void;
   setIsMuted: (isMuted: boolean) => void;
   setPosition: (position: number) => Promise<any>;
@@ -117,6 +118,26 @@ export default function Player(props: Props) {
       );
     }
     return null;
+  };
+
+  const _renderReplaceButton = () => {
+    return (
+      <View>
+        <TouchableOpacity style={styles.button} onPress={props.replace}>
+          <Text
+            style={{
+              textAlign: 'center',
+              fontWeight: 'bold',
+              marginLeft: 8,
+              fontSize: 12,
+              height: 36,
+              color: Colors.tintColor,
+            }}>
+            Replace Source
+          </Text>
+        </TouchableOpacity>
+      </View>
+    );
   };
 
   const _renderAuxiliaryButton = ({
@@ -253,6 +274,7 @@ export default function Player(props: Props) {
             active: false,
           })}
       </View>
+      {_renderReplaceButton()}
       {_maybeRenderErrorOverlay()}
     </View>
   );