+++ packages/expo-audio/ios/AudioPlayer.swift
@@ -25,6 +25,7 @@ public class AudioPlayer: SharedRef<AVPlayer> {
   private var audioProcessor: AudioTapProcessor?
   private var samplingEnabled = false
   private var tapInstalled = false
+  private var shouldInstallAudioTap = false
 
   private var duration: Double {
     (ref.currentItem?.duration.seconds ?? 0.0) * 1000
@@ -93,17 +94,40 @@ public class AudioPlayer: SharedRef<AVPlayer> {
 
   private func setupPublisher() {
     ref.publisher(for: \.currentItem?.status)
-      .sink { status in
-        guard let status else {
+      .sink { [weak self] status in
+        guard let self, let status else {
           return
         }
         if status == .readyToPlay {
           self.updateStatus(with: [
             "isLoaded": true
           ])
+          // We can't add the audio tap until the asset has loaded, otherwise the asset track will be empty.
+          // This is particularly important after replacing the audio source
+          if shouldInstallAudioTap {
+            setSamplingEnabled(enabled: shouldInstallAudioTap)
+            shouldInstallAudioTap = false
+          }
         }
       }
       .store(in: &cancellables)
+  }
+
+  func replaceCurrentSource(source: AudioSource) {
+    let wasPlaying = ref.timeControlStatus == .playing
+    let wasSamplingEnabled = samplingEnabled
+    ref.pause()
+
+    // Remove the audio tap if it is active
+    if samplingEnabled {
+      setSamplingEnabled(enabled: false)
+    }
+    ref.replaceCurrentItem(with: AudioUtils.createAVPlayerItem(from: source))
+    shouldInstallAudioTap = wasSamplingEnabled
+
+    if wasPlaying {
+      ref.play()
+    }
   }
 
   private func playerIsBuffering() -> Bool {
@@ -122,8 +146,8 @@ public class AudioPlayer: SharedRef<AVPlayer> {
   }
 
   private func installTap() {
-    if let item = ref.currentItem, !tapInstalled {
-      audioProcessor = AudioTapProcessor(playerItem: item)
+    if !tapInstalled {
+      audioProcessor = AudioTapProcessor(player: ref)
       tapInstalled = audioProcessor?.installTap() ?? false
       audioProcessor?.sampleBufferCallback = { [weak self] buffer, frameCount, timestamp in
         guard let self,